Up to C_Collider.hpp and cpp: cleaned up;

+++ MyGameEngine/Animation.hpp
@@ -22,31 +22,22 @@ class Animation {
 public:
     Animation(FacingDirection direction);
 
-    void AddFrame(int textureID, int x, int y,
-        int width, int height, float frameTime);
-
+    void AddFrame(int textureID, int x, int y, int width, 
+        int height, float frameTime);
     const FrameData* GetCurrentFrame() const;
-
     bool UpdateFrame(float deltaTime);
 
     void Reset();
 
     void SetDirection(FacingDirection direction);
-
     FacingDirection GetDirection() const;
 
 private:
     void IncrementFrame();
 
-    // Stores all frames for our animation.
     std::vector<FrameData> frames;
-
-    // Current frame.
-    int currentFrameIndex;
-
-    // We use this to decide when to transition to the next frame.
+    size_t currentFrameIndex;
     float currentFrameTime;
-
     FacingDirection direction;
 };
